requirement.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
pdfplumber==0.10.3
PyMuPDF==1.23.8
requests==2.31.0
openpyxl==3.1.2
fpdf2==2.7.7
python-dateutil==2.8.2